Tweak glob and switch threshold direction for labels
glob.glob with the given pattern was not working on my system, so I changed to
pathlib.  I was only getting the outlines rather than solid polygons for my
label rasters, from the XYZ tiles I exported from QGIS. I found it worked if I
did black on a white background and reversed the threshold.  This probably
isn't the ideal solution.