Change intersection protocol.
Originally this used the the logical intersection of the building shapes and
the object shapes. This means that if a building has a small amount of green
roof then the whole building geometry is returned. Instead, I wanted to return
just the part that is the green roof.  You could skip this step entirely, but
it would not exclude false positives where an identified object does not
intersect with a building geometry.